Inverts identation icons (which where wrong)
Actually the "outdent" button uses icon-indent-right, which clearly represents an action of pushing the text to the right side. The same goes with the "ident" button that has an icon representing the opposite action.
If you check the official demo of bootstrap-wysihtml5 (http://jhollingworth.github.io/bootstrap-wysihtml5/) you'll see that the buttons are correct.